requirements.txt
google-cloud-compute==1.15.0
google-cloud-storage==2.10.0
pandas==2.0.3
openpyxl==3.1.2